more Rules and Scoreboard changes
Made enhancements to the rules and scoreboard classes.  

Rules:  Added several functions to determine if Ones thru Sixes are valid.  Also added all the calculations for the dice results that can be returned to the Scoreboard class.
Scoreboard:  Added methods to reset and set a value from the playing board.